chore: break down components in different
